+++ src/main/java/module-info.java
@@ -22,29 +22,23 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.data.identifiable;
-
-/**
- * The interface {@link IdGenerator} provides a method for get the next and the current id object
- *
- * @param <T>
- *            the generic type of the id
- */
-public interface IdGenerator<T>
-{
-
-	/**
-	 * Gets the next id
-	 * 
-	 * @return the next id
-	 */
-	T getNextId();
-
-	/**
-	 * Gets the current id
-	 *
-	 * @return the current id
-	 */
-	T getCurrentId();
-
+module data.api {
+	exports io.github.astrapi69.data.accomplishable;
+	exports io.github.astrapi69.data.activatable;
+	exports io.github.astrapi69.data.auditable;
+	exports io.github.astrapi69.data.creatable;
+	exports io.github.astrapi69.data.deletable;
+	exports io.github.astrapi69.data.enumtype;
+	exports io.github.astrapi69.data.identifiable;
+	exports io.github.astrapi69.data.modifiable;
+	exports io.github.astrapi69.data.nameable;
+	exports io.github.astrapi69.data.pairable;
+	exports io.github.astrapi69.data.processable;
+	exports io.github.astrapi69.data.traceable;
+	exports io.github.astrapi69.data.treeable;
+	exports io.github.astrapi69.data.validatable;
+	exports io.github.astrapi69.data.validatable.versionable;
+	exports io.github.astrapi69.data.verifiable;
+	exports io.github.astrapi69.data.versionable;
+	exports io.github.astrapi69.data.visibility;
 }